refactor: trait-manager inside pf2 and lazy-loading only on pf2e system, removed deprecated setting, added jsconfig and updated related git workflows, formatting
