+++ jsconfig.json
@@ -1,0 +1,6 @@
+{
+	"compilerOptions": {
+		"checkJs": true,
+		"strict": false
+	}
+}
