+++ main.js
@@ -1,4 +1,10 @@
-import './module/settings.js';
 import './module/menu.js';
+import './module/settings.js';
 import './module/token.js';
 import('./module/encounter.js');
+
+Hooks.on('init', () => {
+	if (game.system.id === 'pf2e') {
+		import('./module/pf2/trait-manager.js');
+	}
+});
