+++ module/global.d.ts
@@ -1,0 +1,7 @@
+declare var foundry: any;
+declare var Hooks: any;
+declare var game: any;
+declare var ui: any;
+declare var CONST: any;
+declare var saveDataToFile: any;
+declare var Playlist: any;
