+++ module/transfer.js
@@ -1,5 +1,5 @@
-import { MODULE_ID, getSetting, setSetting } from './settings.js';
 import { parseMusic } from './music-manager.js';
+import { MODULE_ID, getSetting, setSetting } from './settings.js';
 
 /* -------------------------------------------- */
 /*  Export                                      */
@@ -26,7 +26,7 @@ export async function exportMusicConfig() {
 	// Resolve trait rules to names so they survive across worlds.
 	const resolvedTraitRules = traitRules.map((rule) => {
 		const sound = parseMusic(rule.music);
-		const playlist = 'error' in sound ? null : sound.parent ?? sound;
+		const playlist = 'error' in sound ? null : (sound.parent ?? sound);
 		const track = playlist && sound !== playlist ? sound : null;
 		return {
 			trait: rule.trait,
@@ -62,6 +62,7 @@ export function importMusicConfig() {
 
 	input.addEventListener('change', async (ev) => {
 		document.body.removeChild(input);
+		// @ts-ignore
 		const file = ev.target.files[0];
 		if (!file) return;
 		const text = await file.text();
@@ -109,13 +110,15 @@ async function applyImport(data) {
 		for (const soundData of playlistData.sounds) {
 			const existing = playlist.sounds.contents.find((s) => s.name === soundData.name);
 			if (!existing) {
-				await playlist.createEmbeddedDocuments('PlaylistSound', [{
-					name: soundData.name,
-					path: soundData.path,
-					volume: soundData.volume ?? 0.8,
-					repeat: soundData.repeat ?? true,
-					streaming: soundData.streaming ?? false,
-				}]);
+				await playlist.createEmbeddedDocuments('PlaylistSound', [
+					{
+						name: soundData.name,
+						path: soundData.path,
+						volume: soundData.volume ?? 0.8,
+						repeat: soundData.repeat ?? true,
+						streaming: soundData.streaming ?? false,
+					},
+				]);
 			} else {
 				await existing.update({ path: soundData.path });
 			}
@@ -126,22 +129,20 @@ async function applyImport(data) {
 
 	// Resolve trait rules using the map we built during import.
 	if (data.traitRules?.length) {
-		const resolvedRules = data.traitRules.map((rule) => {
-			const playlist = playlistMap.get(rule.playlistName);
-			const track = rule.trackName
-				? playlist?.sounds.contents.find((s) => s.name === rule.trackName)
-				: null;
-			const music = track
-				? (playlist.id + '.' + track.id)
-				: playlist?.id ?? '';
-			return {
-				trait: rule.trait,
-				priority: rule.priority,
-				playlistId: playlist?.id ?? '',
-				trackId: track?.id ?? '',
-				music,
-			};
-		}).filter((r) => r.music);
+		const resolvedRules = data.traitRules
+			.map((rule) => {
+				const playlist = playlistMap.get(rule.playlistName);
+				const track = rule.trackName ? playlist?.sounds.contents.find((s) => s.name === rule.trackName) : null;
+				const music = track ? playlist.id + '.' + track.id : (playlist?.id ?? '');
+				return {
+					trait: rule.trait,
+					priority: rule.priority,
+					playlistId: playlist?.id ?? '',
+					trackId: track?.id ?? '',
+					music,
+				};
+			})
+			.filter((r) => r.music);
 		await setSetting('traitRules', resolvedRules);
 	}
